Add caddy support (#1)
* Add Caddy server integration with mesi middleware

Introduced a Caddyfile and Makefile for building Caddy with the go-mesi plugin. The Caddyfile sets up a local server configuration. The changes also include dependency updates in the go.sum file to support this integration.

* Add mESI middleware support for Caddy

Introduced an implementation of Edge Side Includes (ESI) middleware for the Caddy server. Updated the build process, created a new README with setup and configuration instructions, and corrected the handler directive registration for "mesi" middleware.

* Fix typo and add Caddy plugin link in README checklist

Corrected an extra dash in the Nginx plugin entry for consistency. Added a link to the installation and configuration guide for the Caddy plugin. Improved clarity and completeness of the plugin checklist.